Fix duplicate property label
`EconomicActivity` inherits the `Interval:description` property (labeled "Description") and also defines its own `goodsDescription` property (also labeled "Description"). This leads to confusion when displaying lists of properties in user interfaces.

We may want to reconsider if a separate `goodsDescription` property is necessary. However, this would be a breaking change, so for now I suggest we update the label to be unique.