Display inheritance graph in model explorer (#1114)
* Add basic inheritance graph for schema docs

* Fix schema explorer TOC

* Add loading indicator

* Make use of Mermaid’s own theming

* Refactor: Make inheritance graph reusable

For example to display a full graph of the model on the explorer index page…

* Implement workaround for Astro bug

This is a workaround due to an Astro bug. The bug has been fixed, but we're still on an older Astro version, and using a different file extension works fine for now.

* Improve dark mode

* Adjust edge/arrow styles

* Use schema name (instead of label) in model explorer

Most people using the model explorer will probably search for the schema name as that is what they use when creating mappings, writing crawlers, debugging problems with Aleph etc. The schema label is still visible on the index page, but we now use the name consistently in headings, page titles, navigation links etc.