add parent entity reference property to RealEstate
RealEstate objects can often have children, such as a building has
individual apartments. It often happens that modeling this relationship
is important for investigations. This commit and future pull request
addess the ability for a RealEstate entity to have a parent so
journalists can see subunits of a building, or which apartments/subunits
are associated with a single parent unit.