Optionally disable value cleaning in mappings
Sometimes it’s necessary to take a pragmatic approach to data cleaning. Either because the amount of work required to properly clean everything is simply too high and it’s preferred to work with data that’s a little messy instead of having no data at all. Or because the source data is ambiguous. For example, FollowTheMoney requires phone numbers to specify a country code, but the source data may contain phone numbers without a country code and there is no way to unambiguously derive the country code.

This change allows mapping developers to optionally disable cleaning for individual properties in a mapping. For example, the following example disables cleaning only for the mapped phone property. This allows mapping a value like `(567) 992-1234` to the phone property (which would be removed by default because it’s missing a country code).

```yml
entities:
  person:
    schema: Person
    properties:
      name:
        column: name
      phone:
        column: phone
        clean: false
```