Improve logging when skipping values or entities in mappings (#1425)
* Return the (cleaned) value added to a proxy

This is a small change that allows callers of the `unsafe_add` method to determine if a value has been removed as part of value cleaning. This is helpful to be able to log/warn when values are removed.

* Log warning when discarding unclean values

Right now, mappings silently discard unclean values. This makes it difficult to debug mapping issues during development as well as to monitor mappings running in production.

* Log warning when entities are skipped during mappings