Always use an array for extends and edge.caption
`extends and `edge.caption` was sometimes (not always) a string value
when there's just one entry, instead of an array wiht one entry.

This is kind of annoying to parse and deal with; e.g. in Go I need to
have a custom type with UnmarshalYAML(). Not hugely complex, but 17
lines of code I could do without. And even in more dynmic languages like
Python you'll need an "if" to deal with it.

So just change it to consistently be an array.