Add CLI and mappings docs (previously part of the Aleph docs)
The documentation for the ftm CLI and mappings where previously part of the [Aleph docs](https://docs.aleph.occrp.org/developers). We recently restructured the Aleph docs. As part of that, we remove the ftm CLI and mappings docs. I’m moving them to the FollowTheMoney docs as they are helpful resources.

I’ve only removed a few references to Aleph that would’ve been confusing otherwise and fixed a few links, but apart from that the contents are mostly unedited.